+++ meta-lottie-compose/src/main/java/com/stash/metalottie/compose/ColorType.kt
@@ -1,0 +1,6 @@
+package com.stash.metalottie.compose
+
+enum class ColorType {
+    STROKE,
+    FILL
+}