+++ meta-lottie-compose/src/main/java/com/stash/metalottie/compose/ThemeColorProvider.kt
@@ -1,0 +1,8 @@
+package com.stash.metalottie.compose
+
+import androidx.compose.runtime.Composable
+
+interface ThemeColorProvider {
+    @Composable
+    fun getValue(identifier: ColorValueIdentifier): Int?
+}